Fix warnings in trview.app

+++ trlevel/ILevel.h
@@ -55,7 +55,7 @@ namespace trlevel
 
         // Get the room at the specified index.
         // Returns: The room.
-        virtual tr3_room get_room(uint16_t index) const = 0;
+        virtual tr3_room get_room(uint32_t index) const = 0;
 
         // Get the number of object textures in the level.
         // Returns: The number of object textures.
